Safari-609.1.11.1/LayoutTests/http/tests/security/contentSecurityPolicy/script-redirect-blocked-expected.txt

CONSOLE MESSAGE: Refused to load http://localhost:8000/security/contentSecurityPolicy/resources/alert-fail.js because it does not appear in the script-src directive of the Content Security Policy.
ALERT: PASS
